Task dates: parse and store date and time
Currently, dates are stored as Strings.

This makes it harder to process or display
dates in a consistent format.

Let's parse Deadline and Event date and time
inputs into actual date and time that
can be understood by the chatbot.

As an aside, add a new 'reset' command to
clear all tasks from the list.

Using LocalDateTime instead of Strings enables
future features such as listing tasks by date
or sorting tasks chronologically.